DB based RBAC, most Access Service methods are now deprecated

+++ frontend/controllers/UsersController.php
@@ -24,6 +24,7 @@ use app\components\RequestType;
 
 class UsersController extends Controller
 {
+
     public function behaviors()
     {
         return [
@@ -36,23 +37,23 @@ class UsersController extends Controller
                     ],
                 ],
             ],
-            'verbs' => [
-                'class' => VerbFilter::className(),
+            'verbs'  => [
+                'class'   => VerbFilter::className(),
                 'actions' => [
                     'logout' => ['post'],
                 ],
             ],
         ];
     }
-    
+
     public function actions()
     {
         return [
-            'error' => [
+            'error'   => [
                 'class' => 'yii\web\ErrorAction',
             ],
             'captcha' => [
-                'class' => 'yii\captcha\CaptchaAction',
+                'class'           => 'yii\captcha\CaptchaAction',
                 'fixedVerifyCode' => YII_ENV_TEST ? 'testme' : null,
             ],
         ];
@@ -62,7 +63,7 @@ class UsersController extends Controller
     {
         /////////////////////////--- Profile Infos ---//////////////////////////
         $id = UserService::getUserIdByName($uname);
-        if($id === false)
+        if ($id === false)
         {
             throw new \yii\web\NotFoundHttpException("User cannot be found");
         }
@@ -71,74 +72,70 @@ class UsersController extends Controller
         {
             return $this->redirect('/profile');
         }
-        if (Yii::$app->request->post())
+        if (Yii::$app->request->isPost)
         {
-            if (AccessService::check(Permission::ManageUserRelations))
+            if (Yii::$app->user->can('relations-manage-own'))
             {
                 $request = Yii::$app->request;
-                if (!is_null($request->post('follow-btn')))
+                if (!is_null($request->post('follow-btn')) && Yii::$app->user->can('relations-follow'))
                 {
                     RelationService::setRelation($myId, $id, RelationType::Follower);
                 }
-                if (!is_null($request->post('friend-btn')))
+                if (!is_null($request->post('friend-btn')) && Yii::$app->user->can('relations-friend'))
                 {
                     RequestService::createRequest($myId, $id, RequestType::FriendRequest, date('Y-m-d H:i:s')); //to tutaj
                 }
-                if (!is_null($request->post('unfriend-btn')))
+
+                if (!is_null($request->post('unfriend-btn')) && Yii::$app->user->can('relations-friend'))
                 {
                     RelationService::removeRelation($myId, $id, RelationType::Friend);
                 }
-                if (!is_null($request->post('unfollow-btn')))
+                if (!is_null($request->post('unfollow-btn')) && Yii::$app->user->can('relations-follow'))
                 {
                     RelationService::removeRelation($myId, $id, RelationType::Follower);
+                }
+
+                if (!is_null(Yii::$app->request->post('type')))
+                {
+                    switch (Yii::$app->request->post('type'))
+                    {
+                        case 'newpost':
+                            PostsService::createPost($id, Yii::$app->request->post('inputText'));
+                            break;
+
+                        case 'newcomment':
+                            PostsService::createComment(Yii::$app->request->post('post_id'), Yii::$app->request->post('inputText'));
+                            break;
+                    }
                 }
             }
             else
             {
-                $this->redirect("intouch/accessdenied");
+                $this->redirect(["intouch/accessdenied"]);
             }
         }
-        
-        
-        
+
+
         $education = UserService::getUserEducation($id);
-        $about = UserService::getUserAbout($id);
-        $city = UserService::getUserCity($id);
-        $birth = UserService::getBirthDate($id);
-        $name = UserService::getName($id);
-        $surname = UserService::getSurname($id);
-        if (strlen($name) == 0 || strlen($surname) == 0)
+        $about     = UserService::getUserAbout($id);
+        $city      = UserService::getUserCity($id);
+        $birth     = UserService::getBirthDate($id);
+        $name      = UserService::getName($id);
+        $surname   = UserService::getSurname($id);
+        if (strlen($name) == 0 || strlen($surname) == 0)                //TODO Move it to view
         {
-            $name = "Dane nie uzupełnione";
+            $name    = "Dane nie uzupełnione";
             $surname = "";
         }
-        $email = UserService::getEmail($id);
-        $followers = count(RelationService::getUsersWhoFollowMe($id));
-        $following = count(RelationService::getUsersWhoIFollow($id));
-        $friends = count(RelationService::getFriendsList($id));
-        $photo = PhotoService::getProfilePhoto($id, true, true);
+        $email         = UserService::getEmail($id);
+        $followers     = count(RelationService::getUsersWhoFollowMe($id));
+        $following     = count(RelationService::getUsersWhoIFollow($id));
+        $friends       = count(RelationService::getFriendsList($id));
+        $photo         = PhotoService::getProfilePhoto($id, true, true);
         /////$$$$$ FORMS $$$$$//////////////////////////////////////////////////
-        
-        if (Yii::$app->request->isPost)
-        {
-            if (!is_null(Yii::$app->request->post('type')))
-            {
-                switch (Yii::$app->request->post('type'))
-                {
-                    case 'newpost':
-                        PostsService::createPost($id, Yii::$app->request->post('inputText'));
-                        break;
-
-                    case 'newcomment':
-                        PostsService::createComment(Yii::$app->request->post('post_id'), Yii::$app->request->post('inputText'));
-                        break;
-                }
-            }
-        }
-
         ////////////////////////////--- Other stuff ---/////////////////////////
         $UserRelations = RelationService::getRelations($myId, $id);
-        $isFriend = $UserRelations[RelationType::Friend];
+        $isFriend      = $UserRelations[RelationType::Friend];
         if (!$isFriend)
         {
             if (RequestService::isRequestBetween($id, $myId, RequestType::FriendRequest))
@@ -146,39 +143,39 @@ class UsersController extends Controller
                 $isFriend = "Friend Request Sent";
             }
         }
-        $IFollow = $UserRelations[RelationType::Follower];
-        $uname = UserService::getUserName($id);
+        $IFollow      = $UserRelations[RelationType::Follower];
+        $uname        = UserService::getUserName($id);
         //***Do not add anything new below this line (except for the render)****
         $this->getUserData($id);
         $this->layout = 'logged';
-        $posts = PostsService::getPosts($id);
-        $shared = [
-            'name' => $name,
-            'surname' => $surname,
-            'email' => $email,
-            'education' => $education,
-            'about' => $about,
-            'city' => $city,
-            'birth' => $birth,
-            'followers' => $followers,
-            'following' => $following,
-            'friends' => $friends,
-            'UserFollowState' => $IFollow,
+        $posts        = PostsService::getPosts($id);
+        $shared       = [
+            'name'                => $name,
+            'surname'             => $surname,
+            'email'               => $email,
+            'education'           => $education,
+            'about'               => $about,
+            'city'                => $city,
+            'birth'               => $birth,
+            'followers'           => $followers,
+            'following'           => $following,
+            'friends'             => $friends,
+            'UserFollowState'     => $IFollow,
             'UserFriendshipState' => $isFriend,
-            'UserName' => $uname,
-            'UserProfilePhoto' => $photo,
-            'id' => $id,
-            'posts' => $posts,
-            'photo' => $photo,
-            'myId' => $myId,
-            'myUname' => UserService::getUserName($myId),
+            'UserName'            => $uname,
+            'UserProfilePhoto'    => $photo,
+            'id'                  => $id,
+            'posts'               => $posts,
+            'photo'               => $photo,
+            'myId'                => $myId,
+            'myUname'             => UserService::getUserName($myId),
         ];
-        
+
         $this->getUserData();
         $this->layout = "logged";
         return $this->render('view', $shared);
     }
-    
+
     private function getUserData()
     {
         $id = Yii::$app->user->getId();
@@ -187,19 +184,19 @@ class UsersController extends Controller
 
         if (is_string($photo))
         {
-            $location = "@web/dist/content/images/";
+            $location                               = "@web/dist/content/images/";
             //TODO set chmod for that directory(php init)
             $this->view->params['userProfilePhoto'] = $location . $photo;
         }
         else
         {
-            $location = "@web/dist/img/guest.png";
+            $location                               = "@web/dist/img/guest.png";
             //TODO add that file
             $this->view->params['userProfilePhoto'] = $location;
         }
 
-        $userinfo = array();
-        $userinfo['user_name'] = UserService::getName($id);
+        $userinfo                 = array();
+        $userinfo['user_name']    = UserService::getName($id);
         $userinfo['user_surname'] = UserService::getSurname($id);
         if ($userinfo['user_name'] == false)
         {
@@ -213,9 +210,9 @@ class UsersController extends Controller
         $this->view->params['userInfo'] = $userinfo;
         ////////////////////////////////////////////////////// request service
 
-        $notification = RequestService::getMyRequests($id);
-        $tablelength = count($notification);
-        $this->view->params['notification_data'] = $notification;
+        $notification                             = RequestService::getMyRequests($id);
+        $tablelength                              = count($notification);
+        $this->view->params['notification_data']  = $notification;
         $this->view->params['notification_count'] = $tablelength;
     }
